+++ strict-mvar/test/Main.hs
@@ -1,0 +1,7 @@
+module Main where
+
+import qualified Test.Control.Concurrent.Class.MonadMVar.Strict.Checked as Checked
+import           Test.Tasty
+
+main :: IO ()
+main = defaultMain Checked.tests